pyproject.toml
[project]
name = "document-network-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi>=0.135.1",
    "markdown>=3.10.2",
    "pypdf>=6.7.5",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
